feat: add Set as an alternate storage/search mechanism
Sets can be generally assumed to be implemented as a hash table
and can check keys in constant time, quicker than the tree. There
are trade-offs but without more information the Set is much simpler
to implement/maintain given it's a built-in. Made the Trie's API
match the Set's and added a switch to use one or the other--just
for fun.